refactor: Migrate tool management to Homebrew Brewfiles.
* Migrated development tool management from Go module-based (`go.tools.mod`) and MISE configurations to Homebrew Brewfiles for improved platform compatibility and easier maintenance.
* Added Brewfile specifications for Go, standard, and comprehensive tool installations with versioning and documentation.
* Removed legacy Go tool installation scripts and MISE configuration files in favor of declarative Brewfile approach.
* Created `annotate-brewfiles.sh` utility to automatically annotate Brewfiles with upstream package URLs.
* Updated configuration and exclusion lists to include cache directories and build artifacts.
* Modernized Go codebase to use contemporary reflection and utility APIs (`reflect.TypeFor`, `maps.Copy`, `slices.Contains`).
* Enhanced development linter configuration with additional Go tools (`go-fix`, `unconvert`) for improved code analysis.

+++ cmd/console.go
@@ -326,12 +326,11 @@ func minMaxRows[T any](rows []T) int {
 	return int(
 		math.Min(
 			10,
-			float64(
-				math.Max(
-					5,
-					float64(
-						len(rows),
-					),
+
+			math.Max(
+				5,
+				float64(
+					len(rows),
 				),
 			),
 		),
